requirements.txt
Flask>=2.0
Pillow>=9.0
imagehash>=4.3
